requirements.txt
langchain
langchain-groq
langchain-core
langchain-community
langchain-text-splitters
python-dotenv
pypdf
faiss-cpu
sentence-transformers
rank-bm25
gradio
langchain-ollama